requirements.txt
pytest~=8.1.1
allure-pytest
selene>=2.0.0rc4
Appium-Python-Client~=4.0.0
python-dotenv~=1.0.1
config~=0.5.1
requests~=2.31.0